+++ packages/node/test/feature/module.gwt.js
@@ -1,0 +1,15 @@
+//--------------------------------------------------------
+//-- node/module - Given-When-Then
+//--------------------------------------------------------
+import path from "path";
+import * as gwt from "../../../../test/feature/config.gwt";
+
+const given = { ...gwt.given };
+const when = { ...gwt.when };
+const then = { ...gwt.then };
+
+given.currentRoot = () => {
+	given.root(path.join(__dirname, "..", "..", "dist", "node", "module"));
+};
+
+export { given, when, then };